+++ example/lib/native_app.dart
@@ -76,11 +76,11 @@ class MyApp extends StatelessWidget {
                 builder: (context) {
                   return SingleChildScrollView(
                     child: Padding(
-                      padding: const EdgeInsets.symmetric(horizontal: sbbDefaultSpacing),
+                      padding: const EdgeInsets.symmetric(horizontal: SBBSpacing.medium),
                       child: Column(
                         children: [
                           const Padding(
-                            padding: EdgeInsets.symmetric(vertical: sbbDefaultSpacing),
+                            padding: EdgeInsets.symmetric(vertical: SBBSpacing.medium),
                             child: ThemeModeSegmentedButton(),
                           ),
                           const SBBListHeader('Basics'),
@@ -97,7 +97,7 @@ class MyApp extends StatelessWidget {
                               ).toList(growable: false),
                             ),
                           ),
-                          const SizedBox(height: sbbDefaultSpacing),
+                          const SizedBox(height: SBBSpacing.medium),
                           const SBBListHeader('Elements'),
                           SBBContentBox(
                             child: Column(
@@ -127,7 +127,7 @@ class MyApp extends StatelessWidget {
                               ).toList(growable: false),
                             ),
                           ),
-                          const SizedBox(height: sbbDefaultSpacing),
+                          const SizedBox(height: SBBSpacing.medium),
                           const SBBListHeader('Modules'),
                           SBBContentBox(
                             child: Column(
@@ -149,7 +149,7 @@ class MyApp extends StatelessWidget {
                               ).toList(growable: false),
                             ),
                           ),
-                          const SizedBox(height: sbbDefaultSpacing),
+                          const SizedBox(height: SBBSpacing.medium),
                         ],
                       ),
                     ),