+++ example/lib/pages/segmented_button_page.dart
@@ -1,5 +1,5 @@
-import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 import 'package:flutter/widgets.dart';
+import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 
 import '../native_app.dart';
 
@@ -20,10 +20,10 @@ class SegmentedButtonPageState extends State<SegmentedButtonPage> {
   Widget build(BuildContext context) {
     return ListView(
       children: <Widget>[
-        const Padding(padding: EdgeInsets.all(sbbDefaultSpacing), child: ThemeModeSegmentedButton()),
+        const Padding(padding: EdgeInsets.all(SBBSpacing.medium), child: ThemeModeSegmentedButton()),
         const SBBListHeader('Default (colors based on theme)'),
         Padding(
-          padding: const EdgeInsets.symmetric(horizontal: sbbDefaultSpacing),
+          padding: const EdgeInsets.symmetric(horizontal: SBBSpacing.medium),
           child: SBBSegmentedButton(
             values: ['Option 1', 'Option 2', 'Option 3'],
             selectedIndexChanged: (value) => setState(() => _selectedStateIndex1 = value),
@@ -31,7 +31,7 @@ class SegmentedButtonPageState extends State<SegmentedButtonPage> {
           ),
         ),
         Padding(
-          padding: const EdgeInsets.all(sbbDefaultSpacing),
+          padding: const EdgeInsets.all(SBBSpacing.medium),
           child: SBBSegmentedButton.icon(
             icons: {
               SBBIcons.microscooter_profile_small: 'Microscooter',
@@ -44,7 +44,7 @@ class SegmentedButtonPageState extends State<SegmentedButtonPage> {
         ),
         const SBBListHeader('Red'),
         Padding(
-          padding: const EdgeInsets.symmetric(horizontal: sbbDefaultSpacing),
+          padding: const EdgeInsets.symmetric(horizontal: SBBSpacing.medium),
           child: SBBSegmentedButton.redText(
             values: ['Option 1', 'Option 2', 'Option 3'],
             selectedIndexChanged: (value) => setState(() => _selectedStateIndex3 = value),
@@ -52,7 +52,7 @@ class SegmentedButtonPageState extends State<SegmentedButtonPage> {
           ),
         ),
         Padding(
-          padding: const EdgeInsets.all(sbbDefaultSpacing),
+          padding: const EdgeInsets.all(SBBSpacing.medium),
           child: SBBSegmentedButton.redIcon(
             icons: {
               SBBIcons.microscooter_profile_small: 'Microscooter',