+++ example/lib/pages/slider_page.dart
@@ -17,48 +17,74 @@ class SliderPageState extends State<SliderPage> {
   double _iconlessSliderValue = 50.0;
   final double _disabledSliderValue = 50.0;
 
+  int _selectedStateIndex = 0;
+
+  bool get _isEnabled => _selectedStateIndex == 0;
+
   @override
   Widget build(BuildContext context) {
-    return ListView(
-      padding: const EdgeInsets.all(sbbDefaultSpacing),
-      children: [
-        const ThemeModeSegmentedButton(),
-        const SizedBox(height: sbbDefaultSpacing),
-        const SBBListHeader('Default'),
-        SBBContentBox(
-          margin: const EdgeInsets.symmetric(horizontal: sbbDefaultSpacing),
-          padding: const EdgeInsets.all(sbbDefaultSpacing),
-          child: SBBSlider(
-            onChanged: (value) {
-              setState(() {
-                _defaultSliderValue = value;
-              });
-            },
-            value: _defaultSliderValue,
-            max: _sliderMaxValue,
+    return CustomScrollView(
+      slivers: [
+        SBBSliverHeaderbox.custom(
+          child: Column(
+            spacing: SBBSpacing.small,
+            children: [
+              const ThemeModeSegmentedButton(),
+              SBBSegmentedButton.text(
+                values: ['Enabled', 'Disabled'],
+                selectedStateIndex: _selectedStateIndex,
+                selectedIndexChanged: (val) {
+                  setState(() {
+                    _selectedStateIndex = val;
+                  });
+                },
+              ),
+            ],
           ),
         ),
-        const SBBListHeader('Without Icons'),
-        SBBContentBox(
-          margin: const EdgeInsets.symmetric(horizontal: sbbDefaultSpacing),
-          padding: const EdgeInsets.all(sbbDefaultSpacing),
-          child: SBBSlider(
-            onChanged: (value) {
-              setState(() {
-                _iconlessSliderValue = value;
-              });
-            },
-            value: _iconlessSliderValue,
-            max: _sliderMaxValue,
-            startIcon: null,
-            endIcon: null,
+        SliverPadding(
+          padding: EdgeInsets.symmetric(
+            vertical: SBBSpacing.xLarge,
+            horizontal: SBBSpacing.xxSmall,
+          ).copyWith(top: SBBSpacing.medium),
+          sliver: SliverList.list(
+            children: [
+              const SBBListHeader('Default'),
+              SBBContentBox(
+                margin: const EdgeInsets.symmetric(horizontal: SBBSpacing.medium),
+                padding: const EdgeInsets.all(SBBSpacing.medium),
+                child: SBBSlider(
+                  onChanged: _isEnabled
+                      ? (value) {
+                          setState(() {
+                            _defaultSliderValue = value;
+                          });
+                        }
+                      : null,
+                  leadingIconData: SBBIcons.walk_slow_small,
+                  trailingIconData: SBBIcons.walk_fast_small,
+                  value: _defaultSliderValue,
+                  max: _sliderMaxValue,
+                ),
+              ),
+              const SBBListHeader('Without Icons'),
+              SBBContentBox(
+                margin: const EdgeInsets.symmetric(horizontal: SBBSpacing.medium),
+                padding: const EdgeInsets.all(SBBSpacing.medium),
+                child: SBBSlider(
+                  onChanged: _isEnabled
+                      ? (value) {
+                          setState(() {
+                            _iconlessSliderValue = value;
+                          });
+                        }
+                      : null,
+                  value: _iconlessSliderValue,
+                  max: _sliderMaxValue,
+                ),
+              ),
+            ],
           ),
-        ),
-        const SBBListHeader('Disabled'),
-        SBBContentBox(
-          margin: const EdgeInsets.symmetric(horizontal: sbbDefaultSpacing),
-          padding: const EdgeInsets.all(sbbDefaultSpacing),
-          child: SBBSlider(onChanged: null, value: _disabledSliderValue, max: _sliderMaxValue),
         ),
       ],
     );