+++ lib/src/headerbox/sbb_headerbox_content.dart
@@ -31,7 +31,7 @@ class DefaultHeaderBoxContent extends StatelessWidget {
                 children: [
                   if (leadingIcon != null) ...[
                     Icon(leadingIcon, size: sbbIconSizeSmall),
-                    SizedBox(width: sbbDefaultSpacing * .5),
+                    SizedBox(width: SBBSpacing.xSmall),
                   ],
                   Expanded(child: Text(title, style: style.titleTextStyle)),
                 ],
@@ -40,7 +40,7 @@ class DefaultHeaderBoxContent extends StatelessWidget {
             ],
           ),
         ),
-        SizedBox(width: sbbDefaultSpacing * .5),
+        SizedBox(width: SBBSpacing.xSmall),
         if (trailingWidget != null) trailingWidget!,
       ],
     );
@@ -70,19 +70,19 @@ class LargeHeaderBoxContent extends StatelessWidget {
       children: [
         if (leadingIcon != null) ...[
           Icon(leadingIcon, size: sbbIconSizeMedium),
-          SizedBox(width: sbbDefaultSpacing * .5),
+          SizedBox(width: SBBSpacing.xSmall),
         ],
         Expanded(
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
-            spacing: sbbDefaultSpacing * .25,
+            spacing: SBBSpacing.xxSmall,
             children: [
               Text(title, style: style.titleTextStyle),
               if (secondaryLabel != null) Text(secondaryLabel!, style: secondaryTextStyle),
             ],
           ),
         ),
-        SizedBox(width: sbbDefaultSpacing * .5),
+        SizedBox(width: SBBSpacing.xSmall),
         if (trailingWidget != null) trailingWidget!,
       ],
     );