+++ lib/src/headerbox/sbb_headerbox_flap.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 
 const _flapBorderRadius = BorderRadius.only(
-  bottomLeft: Radius.circular(sbbDefaultSpacing),
-  bottomRight: Radius.circular(sbbDefaultSpacing),
+  bottomLeft: Radius.circular(SBBSpacing.medium),
+  bottomRight: Radius.circular(SBBSpacing.medium),
 );
 
 const _flapIconSize = 24.0;
@@ -35,10 +35,10 @@ class SBBHeaderboxFlap extends StatelessWidget {
     super.key,
     required this.child,
     this.padding = const EdgeInsets.fromLTRB(
-      sbbDefaultSpacing,
+      SBBSpacing.medium,
       0.0, // margin from [SBBHeaderBox] to allow shadow
-      sbbDefaultSpacing,
-      sbbDefaultSpacing * .5,
+      SBBSpacing.medium,
+      SBBSpacing.xSmall,
     ),
   });
 
@@ -67,7 +67,7 @@ class SBBHeaderboxFlap extends StatelessWidget {
       children: [
         if (leadingIcon != null) ...[
           Icon(leadingIcon, size: _flapIconSize),
-          SizedBox(width: sbbDefaultSpacing * .5),
+          SizedBox(width: SBBSpacing.xSmall),
         ],
         Expanded(
           child: Text(
@@ -78,7 +78,7 @@ class SBBHeaderboxFlap extends StatelessWidget {
           ),
         ),
         if (trailingIcon != null) ...[
-          SizedBox(width: sbbDefaultSpacing * .5),
+          SizedBox(width: SBBSpacing.xSmall),
           Icon(trailingIcon, size: _flapIconSize),
         ],
       ],