+++ lib/src/picker/sbb_time_picker.dart
@@ -90,7 +90,7 @@ class SBBTimePicker extends StatefulWidget {
         mainAxisSize: MainAxisSize.min,
         children: [
           Padding(
-            padding: const EdgeInsets.symmetric(horizontal: sbbDefaultSpacing),
+            padding: const EdgeInsets.symmetric(horizontal: SBBSpacing.medium),
             child: SBBContentBox(
               child: SBBTimePicker(
                 initialTime: initialTime,
@@ -107,7 +107,7 @@ class SBBTimePicker extends StatefulWidget {
             ),
           ),
           Padding(
-            padding: const EdgeInsets.all(sbbDefaultSpacing),
+            padding: const EdgeInsets.all(SBBSpacing.medium),
             child: ListenableBuilder(
               listenable: selectedButtonEnabled,
               builder: (context, _) {