+++ lib/src/tab_bar/sbb_tab_bar.icon_delegate.dart
@@ -30,7 +30,7 @@ class _TabIconDelegate extends MultiChildLayoutDelegate {
     items.mapIndexed((i, e) {
       final tabPosition = tabPositions[i];
       final positionX = gestureInsets.left + maxItemWidth * (i + 0.5) - textSizes[i].width * 0.5;
-      final labelX = min(max(sbbDefaultSpacing, positionX), size.width - textSizes[i].width - sbbDefaultSpacing);
+      final labelX = min(max(SBBSpacing.medium, positionX), size.width - textSizes[i].width - SBBSpacing.medium);
       final position = Offset(labelX, tabPosition.dy + tabSizes[i].height);
       positionChild('${e.id}_label', position);
       return position;