+++ lib/src/theme/sbb_theme.dart
@@ -5,6 +5,7 @@ import 'package:sbb_design_system_mobile/src/checkbox/theme/default_sbb_checkbox
 import 'package:sbb_design_system_mobile/src/chip/theme/default_sbb_chip_theme_data.dart';
 import 'package:sbb_design_system_mobile/src/container/container.dart';
 import 'package:sbb_design_system_mobile/src/radio/theme/default_sbb_radio_theme_data.dart';
+import 'package:sbb_design_system_mobile/src/slider/theme/default_sbb_slider_theme_data.dart';
 import 'package:sbb_design_system_mobile/src/switch/theme/default_sbb_switch_theme_data.dart';
 import 'package:sbb_design_system_mobile/src/stepper/theme/default_sbb_stepper_theme_data.dart';
 
@@ -12,9 +13,6 @@ import '../button/theme/default_button_themes.dart';
 import '../container/theme/default_sbb_content_box_theme_data.dart';
 import '../list_item/theme/default_sbb_list_item_theme_data.dart';
 import '../status/theme/default_sbb_status_theme_data.dart';
-
-@Deprecated('Use SBBSpacing.medium and others instead.')
-const sbbDefaultSpacing = SBBSpacing.medium;
 
 class SBBTheme {
   SBBTheme._();
@@ -34,6 +32,7 @@ class SBBTheme {
     SBBListItemThemeData? listItemTheme,
     SBBRadioThemeData? radioTheme,
     SBBStatusThemeData? statusTheme,
+    SBBSliderThemeData? sliderTheme,
     SBBSwitchThemeData? switchTheme,
     SBBTextTheme? textTheme,
     SBBToastStyle? toastStyle,
@@ -53,6 +52,7 @@ class SBBTheme {
     headerBoxStyle: headerBoxStyle,
     listItemTheme: listItemTheme,
     statusTheme: statusTheme,
+    sliderTheme: sliderTheme,
     switchTheme: switchTheme,
     textTheme: textTheme,
     toastStyle: toastStyle,
@@ -73,6 +73,7 @@ class SBBTheme {
     SBBHeaderBoxStyle? headerBoxStyle,
     SBBListItemThemeData? listItemTheme,
     SBBStatusThemeData? statusTheme,
+    SBBSliderThemeData? sliderTheme,
     SBBSwitchThemeData? switchTheme,
     SBBTextTheme? textTheme,
     SBBToastStyle? toastStyle,
@@ -92,6 +93,7 @@ class SBBTheme {
     headerBoxStyle: headerBoxStyle,
     listItemTheme: listItemTheme,
     statusTheme: statusTheme,
+    sliderTheme: sliderTheme,
     switchTheme: switchTheme,
     textTheme: textTheme,
     toastStyle: toastStyle,
@@ -113,6 +115,7 @@ class SBBTheme {
     SBBListItemThemeData? listItemTheme,
     SBBRadioThemeData? radioTheme,
     SBBStatusThemeData? statusTheme,
+    SBBSliderThemeData? sliderTheme,
     SBBSwitchThemeData? switchTheme,
     SBBTextTheme? textTheme,
     SBBToastStyle? toastStyle,
@@ -157,6 +160,9 @@ class SBBTheme {
     final defaultStatusTheme = DefaultSBBStatusThemeData(baseStyle: mergedBaseStyle);
     final mergedStatusTheme = defaultStatusTheme.merge(statusTheme);
 
+    final defaultSliderTheme = DefaultSBBSliderThemeData(mergedBaseStyle);
+    final mergedSliderTheme = defaultSliderTheme.merge(sliderTheme);
+
     final defaultSwitchTheme = DefaultSBBSwitchThemeData(mergedBaseStyle);
     final mergedSwitchTheme = defaultSwitchTheme.merge(switchTheme);
 
@@ -181,6 +187,7 @@ class SBBTheme {
       listItemTheme: mergedListItemTheme,
       radioTheme: mergedRadioTheme,
       statusTheme: mergedStatusTheme,
+      sliderTheme: mergedSliderTheme,
       switchTheme: mergedSwitchTheme,
       textTheme: mergedTextTheme,
       toastStyle: mergedToastStyle,
@@ -202,6 +209,7 @@ class SBBTheme {
     required SBBListItemThemeData listItemTheme,
     required SBBRadioThemeData radioTheme,
     required SBBStatusThemeData statusTheme,
+    required SBBSliderThemeData sliderTheme,
     required SBBSwitchThemeData switchTheme,
     required SBBTextTheme textTheme,
     required SBBToastStyle toastStyle,
@@ -238,6 +246,7 @@ class SBBTheme {
         listItemTheme,
         radioTheme,
         statusTheme,
+        sliderTheme,
         stepperTheme,
         switchTheme,
         textTheme,