+++ test/sbb_content_box_test.dart
@@ -7,23 +7,23 @@ import 'test_app.dart';
 void main() {
   testWidgets('content_box', (WidgetTester tester) async {
     content(String text) => SizedBox(
-      height: sbbDefaultSpacing * 2,
+      height: SBBSpacing.xLarge,
       width: double.infinity,
       child: Center(child: Text(text)),
     );
 
     final widget = Padding(
-      padding: const EdgeInsets.symmetric(horizontal: sbbDefaultSpacing),
+      padding: const EdgeInsets.symmetric(horizontal: SBBSpacing.medium),
       child: Column(
-        spacing: sbbDefaultSpacing,
+        spacing: SBBSpacing.medium,
         children: [
           SBBContentBox(child: content('Default')),
           SBBContentBox(color: SBBColors.royal, child: content('Different Color')),
           SBBContentBox(
-            padding: EdgeInsets.symmetric(vertical: sbbDefaultSpacing),
+            padding: EdgeInsets.symmetric(vertical: SBBSpacing.medium),
             child: content('Extra padding'),
           ),
-          SBBContentBox(margin: EdgeInsets.all(sbbDefaultSpacing * 4), child: content('Extra margin')),
+          SBBContentBox(margin: EdgeInsets.all(64.0), child: content('Extra margin')),
         ],
       ),
     );